+++ tuple/pair/pair.go
@@ -2,14 +2,14 @@ package pair
 
 // X is a value of the Cartesian cross-product of the given types.
 // Cross-product is represented by an x.
-type X[V1, V2 any] struct {
-	V1 V1
-	V2 V2
+type X[A, B any] struct {
+	First  A
+	Second B
 }
 
-func Of[V, V2 any](v V, v2 V2) X[V, V2] {
-	return X[V, V2]{
-		V1: v,
-		V2: v2,
+func Of[A, B any](a A, b B) X[A, B] {
+	return X[A, B]{
+		First:  a,
+		Second: b,
 	}
 }